Modularize attachment transfer persistence
Separate transfer types, reconciliation, job lifecycle, and completion transactions behind the existing store API.
